split test code
